refactor: move compliance data model into anoma-rm-core (wave 3)
ComplianceInstance, ComplianceWitness (with the pure from_parts
constructor made public), KindTableEntry, the aggregation-guest witness
types, delta_msg/tags, and the kind-table hash move to core.
INITIAL_ROOT becomes a const (same .bss rationale as PADDING_LEAF).

The engine keeps the k256/rand operations as free functions in
compliance::{from_resources, from_resources_with_ephemeral_root,
constrain, delta_projective, kind_table_entry}; the kind-table loader
now reuses core's hash_kind_table_entries instead of its own copy of
the same algorithm. Call sites migrated: compliance guest main,
arm_test_app, compliance_unit, constants, compliance bench.
