+++ arm_circuits/compliance/methods/guest/src/main.rs
@@ -6,7 +6,7 @@ use risc0_zkvm::guest::env;
 pub fn main() {
     let compliance_witness: ComplianceWitness = env::read();
 
-    let compliance_instance = compliance_witness.constrain().unwrap();
+    let compliance_instance = anoma_rm_risc0::compliance::constrain(&compliance_witness).unwrap();
 
     env::commit(&compliance_instance);
 }
