+++ arm_tests/arm_test_app/src/lib.rs
@@ -149,7 +149,7 @@ impl Tester {
         self.populate_created_resources(created_num)?;
 
         init_test_kind_table();
-        let compliance_witness = ComplianceWitness::from_resources(
+        let compliance_witness = anoma_rm_risc0::compliance::from_resources(
             &self.consumed_data[self.current],
             &self.created_resources[self.current],
             global_kind_table().to_vec(),
@@ -682,14 +682,14 @@ fn test_invalid_created_nonce_rejected() {
     tester.set_created_nonce(0, 0, [0xAA; 32]);
 
     init_test_kind_table();
-    let witness = ComplianceWitness::from_resources(
+    let witness = anoma_rm_risc0::compliance::from_resources(
         &tester.consumed_data[0],
         &tester.created_resources[0],
         global_kind_table().to_vec(),
     );
 
     assert!(matches!(
-        witness.constrain(),
+        anoma_rm_risc0::compliance::constrain(&witness),
         Err(ArmError::InvalidResourceNonce)
     ));
 }
